Adding overlay controller, PlayerData parser, and test GUI.

+++ src/mingwMain/kotlin/classes/session/CBApi.kt
@@ -41,7 +41,7 @@ interface ClashBountyApi {
 /**
  * Class object produced by [XrdClashBountyApi]
  */
-class PlayerData(val displayName: String, val steamUserId: Int) {
+class PlayerData(val displayName: String, val steamUserId: Long) {
     var characterId: Byte = 0x0
     var matchesPlayed: Byte = 0x0
     var matchesWon: Byte = 0x0
